% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejemplo.R
\name{divisores}
\alias{divisores}
\title{Divisores}
\usage{
divisores(x)
}
\arguments{
\item{x}{un entero}
}
\description{
Calcula los dividores del numero x.
}
\details{
Por escribir
}
\examples{
divisores(10)
}
