% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejemplo.R
\name{is.perfecto}
\alias{is.perfecto}
\title{Perfecto}
\usage{
is.perfecto(x)
}
\arguments{
\item{x}{un entero}
}
\description{
Verifica si x es o no perfecto.
}
\details{
Por escribir
}
\examples{
is.perfecto(10)
is.perfecto(6)
}
