% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejemplo_nueva_funcion.R
\name{multicolinealidad}
\alias{multicolinealidad}
\title{Multicolinealidad}
\usage{
multicolinealidad(x)
}
\arguments{
\item{x}{un data.frame o matriz}
}
\description{
Calcula el factor de inflacion de varianza de todas las combinaciones posibles
}
\details{
Por escribir
}
\examples{
x<-matrix(rnorm(100*5),nc=5,nr=100)
multicolinealidad(x) 
}
